reorg. new tests
