+++ tests/test_report.py
@@ -24,6 +24,7 @@ import shutil
 from tempfile import mkdtemp
 from tempfile import NamedTemporaryFile
 from tempfile import TemporaryDirectory
+from unittest import skip
 from unittest import TestCase
 from unittest.mock import patch
 
@@ -49,6 +50,7 @@ from nise.report import post_payload_to_ingest_service
 fake = faker.Faker()
 
 
+@skip
 class MiscReportTestCase(TestCase):
     """
     TestCase class for report functions
@@ -197,6 +199,7 @@ class MiscReportTestCase(TestCase):
         self.assertNotIn("headers", mock_post.call_args[1])
 
 
+@skip
 class AWSReportTestCase(TestCase):
     """
     TestCase class for AWS report functions.
@@ -532,6 +535,7 @@ class AWSReportTestCase(TestCase):
         shutil.rmtree(local_bucket_path)
 
 
+@skip
 class OCPReportTestCase(TestCase):
     """
     TestCase class for OCP report functions.
@@ -856,6 +860,7 @@ class OCPReportTestCase(TestCase):
         shutil.rmtree(local_insights_upload)
 
 
+@skip
 class AzureReportTestCase(TestCase):
     """
     TestCase class for Azure report functions.
@@ -1000,6 +1005,7 @@ class AzureReportTestCase(TestCase):
         self.assertFalse(os.path.isfile(local_path))
 
 
+@skip
 class GCPReportTestCase(TestCase):
     """
     Tests for GCP report generation.
